(wip) wip

+++ services/products/models.py
@@ -1,7 +1,8 @@
-from sqlalchemy import Column, ForeignKey, String, Float
+from sqlalchemy import Column, ForeignKey, String, Integer
 from sqlalchemy.ext.declarative import declarative_base
 from sqlalchemy.orm import relationship
 from sqlalchemy.sql import func
+from sqlalchemy.orm import declarative_base
 
 Base  = declarative_base()
 
@@ -10,4 +11,4 @@ class Product(Base):
     id = Column(Integer, primary_key=True, index=True)
     name = Column(String)
     description = Column(String)
-    price = Column(Float)
+    price = Column(Integer)
