+++ services/products/schema.py
@@ -3,7 +3,7 @@ from pydantic import BaseModel
 class ProductSchema(BaseModel):
     name:str
     description:str
-    price:float
+    price:int
 
     class Config:
         orm_mode = True
